Authenticate list with biometrics
